testes finalizados

+++ Back-end/techbridge/src/main/kotlin/com/techbridge/techbridge/dto/InscricaoDTO.kt
@@ -5,6 +5,6 @@ import java.time.LocalDateTime
 data class InscricaoDTO(
     val idInscricao: Long?,
     val idAtivacaoEvento: Long?,
-    val idUsuario: Int?,
+    val idUsuario: Long?,
     val dataInscricao: LocalDateTime
 )
